refactor: restructure project logic heirarchy
The final part for my DDD project structure.
