+++ src/adapter/git.spec.ts
