+++ src/adapter/gpg.spec.ts
